+++ t/00-load.t
@@ -11,7 +11,7 @@ use lib 't/lib';
 
 our $VERSION = '0.001';
 
-const my $TEST_COUNT => 336;
+const my $TEST_COUNT => 339;
 
 plan tests => $TEST_COUNT;
 
@@ -48,6 +48,7 @@ use_ok('GPForum::Command::HypnotoadBenchmark');
 use_ok('GPForum::Command::HypnotoadScaling');
 use_ok('GPForum::Command::OutboxDispatch');
 use_ok('GPForum::Command::ScheduledJobs');
+use_ok('GPForum::Command::StagingDrillAttachments');
 use_ok('GPForum::Command::PlatformCheck');
 use_ok('GPForum::Command::QueryBudget');
 use_ok('GPForum::Query::ReadModel');
@@ -272,6 +273,8 @@ use_ok('GPForum::Service::Operations::BatchPurge');
 use_ok('GPForum::Service::Operations::RetentionStore');
 use_ok('GPForum::Service::Operations::ScheduledJobs');
 use_ok('GPForum::Service::Operations::RunbookValidator');
+use_ok('GPForum::Service::Operations::AttachmentFilesystemDrill');
+use_ok('GPForum::Service::Operations::DeployChecklistDrill');
 use_ok('GPForum::Service::Operations::RuntimeSizing');
 use_ok('GPForum::Service::Portability::ExportBundleBuilder');
 use_ok('GPForum::Service::Portability::ImportJobStore');